Changing sodium to double type in NutritionalValues class

+++ MyHealth.Fitbit.Nutrition/MyHealth.Fitbit.Nutrition/Models/NutritionalValues.cs
@@ -10,6 +10,6 @@ namespace MyHealth.Fitbit.Nutrition.Models
         public double fat { get; set; }
         public double fiber { get; set; }
         public double protein { get; set; }
-        public int sodium { get; set; }
+        public double sodium { get; set; }
     }
 }
